네트워크 기본 세팅 및 Fusion 세팅

+++ program/T3/Assets/DataScript/LevelData.cs
@@ -3,6 +3,6 @@ using UnityEngine;
 [CreateAssetMenu(menuName = "Scriptable/LevelData", fileName = "LevelData")]
 public class LevelData : ScriptableObject
 {
-    public int Cost;
+    public float Cost;
     public int Level;
 }
